feat: add Claude Code collector and provider badges
- ClaudeCodeCollector reads the user's own Claude Code login and
  reports the five-hour and weekly windows; verified against a live
  response
- percent widened to double; Claude reports fractional utilisation
- ProviderNotReadyException carries an instruction the user can act
  on, replacing opaque failures
- brand-coloured monogram badges on cards and in settings, so the
  provider reads at a glance; no third-party logo assets bundled
- document the credential policy: read only, never refresh tokens

+++ src/Manapoint/Collectors/OpenCodeGoUsageParser.cs
@@ -36,7 +36,7 @@ public static class OpenCodeGoUsageParser
 
             windows.Add(new UsageWindow(
                 kind,
-                w.GetProperty("percent").GetInt32(),
+                w.GetProperty("percent").GetDouble(),
                 w.GetProperty("resetsAt").GetDateTimeOffset()));
         }
